move hover button widget into its own file
